Fix libssl/crypto symlinks on Fedora
Fedora does not have the 1.0.0 versions of the .so files, so the
existing symlink step isn't working correctly.  This change looks
for any 1.0.* versions of libssl and libcrypto, so we shouldn't have
to update it for every minor version bump but it also won't
unintentionally pull in an incompatible version.

Change-Id: I8402d05be684337df70d8f48e6d2ac1ecc69eaf7
Related-Bug: #1304629