Drop gearman. (We use geard instead)
This drops the gearman element to avoid confusion with geard.
Initially we tried gearman. It turned out not to be a good
fit because it isn't what openstack-infra uses (they use
geard) instead.

As we already have a working geard element this should help
understanding the CI stack in TripleO less confusing to new
users.

Change-Id: I1df55ff8ab8bf401ac85cdc1a00bdfe5dd391c4f