Fix networking on the control node after reboot
When the control node is rebooted, ethX interface receives
configuration via DHCP, but still is a port of an OVS bridge.
The bridge itself seems to be configured correctly, but there
are no flows in ovs-dpctl output, so the networking is basically
down until you do os-refresh-config.

This is mostly covered by the following commits:
  - t-i-e: Ib775598d3bdd6f1ccecb283f47d10b92aaf0bd35
  - d-i-b: I3166951c6fbb2faec53ffc339edd2fd0c3512300

by making bridge/interface configuration persistent by the means of
network configuration files.

There is still one edge case left. When the node is PXE booted,
"ip=..." command line is supplied to the kernel, and the latter
configures the network interface. On ubuntu this leads to a situation
when the OVS bridge is configured correctly and receives network
configuration over DHCP, but ethX still has an IP address, so
there are two routes to 192.0.2.0 network.

The solution is to remove the ip address from ethX interface right
before the bridge is up. Curiously, this is only needed for Ubuntu.

Closes-Bug: #1304424

Change-Id: Idd6b5b3f98d00286aad512a3b55dfb7fa75a5fb8