Store swift data, db & objects on /mnt filesystem
This commit moves the location of the swift account, container and
object data from the root filesystem (/srv/node) to the /mnt
filesystem. This is done for 2 reasons:
  1. The root filesystem tends to be quite small, swift data can fill
     it quite quickly with nasty consequences (for the OS and swift)
  2. Swift data on the root filesystem does not survive a re-install.
     While swift data on the /mnt filesystems does survive.

This commit creates the /mnt/state/srv/node directory and sets
'devices = /mnt/state/srv/node' in the account, container and object
conf files. It also sets 'path = /mnt/state/srv/node/' in
/etc/rsyncd.conf.

Change-Id: I39f70ba9cf4a8c4c068c95667608afafe700f301