Use Percona Packages for Ubuntu and Debian
Adjust the install of Percona XtraDB Cluster to use Percona Packages
instead of a Percona tarball for Ubuntu and Debian.

This pulls the packages in from the Ubuntu repository. The minimum requirement
for the Percona packages is trusty.

This element currently will not work on Fedora and checks for it but Fedora
uses the Mariadb element to provide a MySQL HA solution instead.

Change-Id: Ie13bf0f7a972ed7f69623c9a661065b46441c3e9