Auto-collect local checks
check_mk supports the option to add checks by dropping executables
into a known directory.

This directory varies on Debian and RHEL-based systems, therefore
this patch allows for users to place these checks into a new
top level directory in the element named "check_mk_checks".

Any script found in here will be deployed to the relevant location
on the target host as part of the image build.

Change-Id: I96312e8b23cb2ff14a19509370a1c48504d73ce8