Ensure rabbitmq can be successfully restarted
We jump through many hoops in order to successfully restart rabbitmq.
This is mostly because we shift the rabbit PID file location, which
means that beam can't be correctly stopped.

Since we're installing static configuration, do that in the image
element rather than at runtime.

Then, ensure that rabbit isn't started until it's actually configured.

We do this in a fashion which is consistent with the operation of
os-svc-enable.

Change-Id: Ifee1dd90eb4db2c29a64864d1cc9fdc9cd3b1bfd