Don't restart mysql during reset-db
A part of reset-db script is cleaning up old files and
restarting DB, this is not ideal in cluster environment
where simple restart is not sufficient - cluster has to
be bootstrapped at first.

Proper mysql startup is handled by a separate orc script.

Cleaning mysql files is not needed too because both percona
and mariadb installations do not pre-create these files.
File ownership is set properly for mysql/mariadb too.

Change-Id: Ida81b53a33bdf0240ccd1432e6ee5da1e7fb0133