Fix missing reference to ceilometer sources.json file
The ceilometer element is trying to install a file that doesn't
exist anymore, which makes the gate fail. This change adds a check
so that we can ignore the file and continue if it doesn't exist
on disk - the gate will succeed, and we should still be compatible
with older versions of ceilometer which have this file.

The file was removed after the Icehouse release which drop v1 api.

Change-Id: I1ec8fbb06e3c26ee1560efc13470e8370149ff89