Add check_mk swift proxy diagnostic
Added a check_mk script that runs healthcheck to ensure that a
proxy node is up and running with proxy services. This diag is
also check_mk compliant (ie. the output matches check_mk desired
format).

Have not added thecheck_mk-agent element to element-deps file as
even though it is required to install the file itself into the build
image, it can be manually added to images that require diagnostics.

Change-Id: If18acfa1b5c967481f59b9f2041d8d7c37db15ca