Name disable-requiretty files uniquely
Updates the nova and ironic elements so that we use unique names
for the pre-install.d scripts to disable the sudo requiretty
feature.

In this commit we also make
elements/ironic/pre-install.d/00-ironic-disable-requiretty
executable (previously it wasn't)

This fixes issues in trying to use the nova/ironic elements on
Fedora 20.

Change-Id: Ic0970eaf3485753b09646b35da59bb43c6acfea7
Closes-bug: #1321849