Add swiftclient library to swift
The swift-dispersion-populate/report tools in swift require
keystone authentication. In order to do this they need to import
the function "get_auth" from swiftclient. Therefore, swiftclient
is a requirement for running these tools and needs to be installed
because, similar to keystoneclient, it is not included in swift's
requirements.txt.

Change-Id: I9d486f8f484e98267caf24720faa49087582b690