Fix l3.filters and dhcp.filters file for tripleo installations
This change addresses a rootwrap filter issue specific to the l3-agent
and dhcp-agent in a tripleo deployment.  The location of the python
executable used in starting processes within a tripleo deployment
may be different than the /usr/bin/python location that is used
for the KillFilter in l3.filters and dhcp.filters.

Closes-Bug: #1272816

Change-Id: I4c78b074e298f49eae559c7ca451fa7f9fec5970