Passthrough-enable for nova-<service>.conf
This permits passhthrough settings to be applied to
nova-<service>.conf. The idea is that under the 'nova' key in the
config we can specify specific service key for defining configuration
for each service.

Here an example of a heat config template that allows configuration for
nova-scheduler service in addition to the general nova config section:
{
  "nova": {
      "scheduler_config": [
            {"section": "DEFAULT",
             "values": [
                {
                   "option": "rpc_thread_pool_size",
                   "value": "2"
                 }
             ]
            }
      ],
      "config": [
            {"section": "DEFAULT",
             "values": [
               {
                   "option": "enable_new_services",
                   "value": "True"
               }
            ]
           }
        ]
    }
}

This depends on https://review.openstack.org/97785
blueprint: passthrough-config

Change-Id: I16ddffb650c66da30652e6abbc0ef5fca378ae70