Add a sysctl-append-value script
The sysctl element works on the assumption that sysctl values are
going to be set by elements that are not used in the same image, and
that therefore, conflicting options would be bad. Some elements would
like to append options to a list, allowing multiple elements to
change the same value.

Change-Id: I4bbfbac716b842210cea2b6af289da78043fd4a7