Add nova url to neutron.conf
Without specifying a nova url in neutron.conf
neutron will attempt to talk to nova on localhost by
default. With the addition of haproxy to tripleO
api services are now proxied using haproxy.  This means
that the default of localhost for nova url is not taking
advantage of the HA endpoint in neutron.conf.
We need to talk to the controller VIP, this change
enables that.

Change-Id: I5eb774e72c101a41cdee887abd19c6628be6df59