Support new virtual_interfaces definition for VIPs
Prior to this change, the keepalived element only allowed for a single
keepalive interface. This could be problematic if, for instance,
keepalive is managing several VIPs which are on different broadcast
domains - for instance, because one vip is on an interface/vlan that's
accessible to the public, while another is on an interface/vlan that's
only accessible within the cloud.

This change allows for a more flexible configuration in which multiple
vrrp instances can be defined in heat- each with their own interface,
router_id, and set of virtual IPs.

This patchset partially completes a spec which will one day live at
https://blueprints.launchpad.net/tripleo/+specs/tripleo-juno-virtual-public-ips
but for now can bee seen in  Id9addc65f0d2ed519ce4b3edbd561ed660a2786e

Removed the check weight configuration as vrrp_sync_groups and and
check_script weights are mutually exclusive configurations. Regardless,
based on
https://github.com/acassen/keepalived/blob/master/doc/keepalived.conf.SYNOPSIS#L103
the default weight is automatically set to 2 if not specified.

Change-Id: I5f04cee6c84ce7ad6e258c489088a47f00a53fda
Co-authored-by: Julia Kreger <juliaashleykreger@gmail.com>