Add swift-ring element
A swift-ring element consists of the basic swift package and the
swift-ring-builder tool. The purpose of this element is exclusively to create
and update swift ring files. The benefit here is that it may be used on an
undercloud or overcloud controller node without having to install the entire
swift service. Also, updates to a ring/creation of a new ring may be
performed on a single, centralised node in the under/overcloud and distributed
to all the swift nodes in a system.

This is not intended to replace anything in the existing swift element. Unlike
the swift element, it does not require any heat-teamplate which is used to
generate swift rings on the node in question (as well as other things) so right
now any node with the swift element must be a part of the swift rings.
Therefore, the swift-ring element could be added to any node in a system
without any other swift services or presence in the rings required.

Change-Id: I018bdb7a8c6f2f7f83a0a7fb6d27332b6ca266a8