ensure-bridge: bring up bridges first on Debian
We should bring up the interfaces on boot in the same order
with which we bring them up in this script.

In this script we write out /etc/network/interfaces and then
bring up the bridges first, and interfaces second. This
allows us to bring up bridges using DHCP with attached ports
and is working quite well on initial deployment.

It has been reported that upon reboot (sometimes?) an Ubuntu
system does not obtain its DHCP ip_address correctly. This
is most likely due to the fact that ifup -a (which is
executed at boot time via networking.conf) brings up
the interfaces *in order* by which they are listed in
/etc/network/interfaces.

As this order does not match the order in this script this
is most likely the cause of the race which sometimes causes
DHCP to fail.

Change-Id: I91628cb78565aec9764e0dcb33c1b1fe165b7c4a
Closes-bug: #1329238