Network init order fixes for VIP
These changes revise the boot order for network intialization such
that the configuration, VIP, HAProxy, and MySQL are available
and running prior to openstack-init starting.

These changes are necessary in order to prevent services in the boot
sequence from failing once the MySQL database URLs and bind-address
are changed to utilize the VIP address.

Service startup changes:

haproxy-nonlocal-bind in configure.d was moved from step 75 to step
20 to execute before haproxy was started so it could could perform
the non-local bind.

haproxy was moved from post-configure.d/ to configure.d/21-haproxy
to allow for HAProxy to bind to the VIP address that keepalived will
need to initalize and support passing connections to MySQL.

neutron-openvswitch-agent in configure.d was moved from starting at
step 80 to starting at step 21 as it creates the bridge interface
that keepalived utilizes, and thus is required for the vip to come
online.

os-apply-config in configure.d was moved from 50 to 20 as the
configuration files must be written out for haproxy and keepalived
to function properly.

This change is depends on I29613bf314fd054d0c82492692d95de2303c64ae

Change-Id: Ia6f26305f8e744e4ff938dff85de1193183ecd8f