Add local_bind flag to my.cf
This change enables the MySQL configuration file's network socket
bind to be changed from 0.0.0.0, to the local-ipv4 address via
the heat template.

    mysql:
      local_bind: true

The absence of the option will result in the bind returning to
the default of 0.0.0.0. This change was implemented this way
so the changes to the heat template could land later.

Change-Id: Iac1274cc52014f25887d696261b32146afc926dd