Create/configure cinder's image_conversion_dir
Updates the cinder element to provide a new top level configuration
for cinder.image_conversion_dir. We are doing this as a top level
config because we also want to create this directory on disk as
well... and using os-apply-config to supply that default makes
sense. (using the pass-thru config would seem to be a bad choice
for this given we need to create a directory if it is set).

The cinder-volume element has been updated to create the
configured image_conversion_dir. If no image_conversion_dir
the default value is created at /mnt/state/var/lib/cinder/conversion.

Change-Id: Ibe8c20b7c815bb84d40a8593e51e82be95e01ec0
Closes-bug: #1340479