Compile and install custom SELinux policies
Provides a way to compile and install custom SELinux policies
to images. Custom policies are placed in a custom-policies
directory. An install.d script will copy the files to
/opt/stack/selinux-policy. When an instance boots, an o-r-c
script will compile the policy source files in that directory
and then load them.

When SElinux is in enforcing mode in the ci, you would need to
create a custom SELinux policy to temporarily fix an upstream
issue to allow ci to pass before that issue has been fixed in the
upstream SELinux policy.

Change-Id: I1cda87e980522f6c668debcf85668bb631d070b3