Move python-memcached install after venv install
When not using common-venv the venv will not exist until after the
horizon install.

We also want to rewrite the manifest after changing the contents of
the venv, to track what is installed.

This is not added to the horizon requirements file as it is an
optional requirement based on whether memcached is being used or not.

Change-Id: I37b2c5bd95b6d6b484986b957e2866c2b57a5b4c
Closes-Bug: #1344014