Adding haproxy reload check
When an initializing controller node restarts haproxy as part of the
initialization sequence, and that node happens to be in control of
the VIP address but is not the bootstrap node, then the initial
cluster initialization sequence can fail.

This commit changes the haproxy script such that if haproxy is
running, then it is reloaded instead of restarted.

Change-Id: I1b35d3ebdb08e175fb0fd58c0c395b7b30103b4b
Closes-bug: #1348931