Use pacemaker to run neutron l3 agent and metadata agent
Currently l3-agent and metadata agent can not run in a/a mode
without loosing data if a node goes down. This patch sets a/p
mode for both services.

Note that this is a short-term solution of providing HA for
the services. Other temporary solutions like running l3-agent in a/a mode
and re-assign routers when a node goes down, might work better but
implementation would be much more complex and it would be still
just a temporary solution until l3-agent supports a/a HA (which is
planned for Juno).

Relies on: Icc97e36a1db198b973041346cf2056f68de661a2
Change-Id: I63f054a8c80f9f676a77341c89e605b5b472d078