Support Debian distro for iptables
We currently use the add-rule script to create iptables rules that allow
the cloud we deployed to function.

These iptable rules are required on RedHat based distros that have a
default deny-all policy; but they're also useful on Debian distros if
the operators turn on a deny-all policy as part of locking down their
environment. It would be useful if these operators could leverage the
work RedHat has done to get the Debian distro working.

This change adds a check for Debian and allows the add-rule script to
run, as a first step towards full support for Debian-based distros.

Also, install.d installs iptables and there is a Debian specific change.

Change-Id: Iea773d37b18c15a417896e93e29bcdc1e20096ac
Closes-Bug: #1351412