Try to start neutron-server first
neutron-l3-agent fails to start if neutron-server does not respond
to an rpc message within one minute.  Because of the order our
neutron scripts run in, this often happens.  To work around
the problem, this change moves the neutron-server script from
the 80 level to 79 so it will be started before the other neutron
services.

Ultimately this needs to be solved in neutron-l3-agent, but right
now this is blocking our CI so we need to address it immediately.

Change-Id: I604adab055fe9c1b0d7bee9cc30ac79afb2d2315
Partial-Bug: 1353953