Set cinder_catalog_info in nova.conf
This commit updates the default nova.conf for TripleO so that
we pin the default cinder_catalog_info to the Cinder V1
API.

Nova recently broken backwards compat for people using the
default (it changed from v1 -> v2 when no config was specified).

As TripleO requires v1 it won't hurt to go ahead and set it
in our default config while we wait for the Nova change to land.

Partial-bug: #1354021

Change-Id: I768c1867f862ff0f10fa98c78907405099c8ff93