Symlink /etc/my.cnf instead of deleting it
All of the existing mariadb elements install from packages. /etc/my.cnf
is a packaged file, and it shouldn't be deleted as rpm does not provide
a way to "mark" a file as intentionally deleted.  Instead of deleting
it, create a symlink to the config we actually use at /etc/mysql/my.cnf.

Change-Id: I3909cd8fbd77a9121f4e3ba4d40eae799207bff9