Secure MySQL clustering with SSL
Accept values through Heat for the contents of an X509 certificate and
key to be used in Galera clustering and write the values into files.
Set the file owner and permissions to only allow mysql access to the
certificate/key pair. Configure Galera clustering to bind to the server
ip:4567 and use the provided certificate/key pair.

This is dependent on the associated change in tripleo-heat-templates
to provide the certificate values.  If values are missing, the
clustering will not be secured.

Turning on/off secure clustering cannot be done without restarting every
node cluster.

Change-Id: I2f3406cd09a2a513347593dfcfd3121d3de096c9