Modify symbolic link creation.
Add a guard so that if /var/log/httpd exist that file is not created
in the /var/log/apache2 directory linking to itself. Also make the
symbolic link creation treat LINK_NAME as a normal file if it is a
symbolic link to a directory. Additionally add a check that if
another element (such as Idced49d2ada450d7948036a9f475c33e29d65712)
has created /var/log/httpd it is not recreated.

Change-Id: I3c08c54c02f2737ee00ac25e34953aa7b3ab546f