Move Cinder NFS backend driver config in cinder-volume-nfs element
This change aims at moving out of cinder/cinder-volume elements
the bits needed to configure the Cinder NFS backend driver and
as such add the nfs-client element which is a dependency for the
NFS backend driver.

To configure Cinder with a pre-existing NFS share, one has to add
cinder-volume-nfs to the image elements and populate the Heat
template with the NFS shares list as per README files.

Change-Id: I79725a19fc61946f9b0f92a0b54c9b3a440b247a