mariadb: call os-svc-enable
When configuring Mariadb we should always auto-start the
Mariadb (or mysql) daemon on startup. This prevents breaking simple
use cases like rebooting nodes, etc.

This patch updates the mariadb-rpm element to autostart mariadb.

The mariadb element is also updated to autostart 'mysql' (which
is the service name that element uses instead).

Change-Id: I02ccead89ce5054557a7037cbf127cbe6e75425b