Add an element to restore ssh keys from /mnt/state
This is intended to work around a race condition where cloud-init looks
for ssh keys on the state partition before it has loaded the state
partition. Since it only finds an empty directory, it goes ahead and
creates new keys, which is undesirable.

This change will restore the keys from /mnt/state when os-refresh-config
is run.

Change-Id: I094d4a8322410a3f2c6d1711c03b998615675aee