Run cinder-volume in active/passive mode for NFS backend
Running multiple cinder-volume instances which access the same storage
is currently not supported and may cause data inconsistency.
This issue is discussed here:
https://bugs.launchpad.net/cinder/+bug/1322190

Running cinder-volume service in active/passive mode seems to be
the only safe way when shared backend storage is used. Also
workaround with setting shared host is needed.

Of course in default no-HA setup each cinder-volume manages its
local volumes and none of the above is needed.

Change-Id: Id8f5c33f63538a9be85a1434f36b0a840e0b8db9