Allow swift "min_part_hours" to be configurable
With swift rings, a "min_part_hours" value is required. This is the
time (in hours) before a specific partition can be moved in succession
(following a rebalance). Currently this value is hard-coded to 1 but
this value is not feasible for a large system and it can often take
several hours for partitions to re-align following a rebalance.

Therefore, this change allows this value to be configurable by adding
the "MIN_PART_HOURS" variable. If this change is accepted, I will add
the relevant parameter in tripleo-heat-templates.

Change-Id: Ie6438386a54794e1ae2f31ad925db01c2c10ed6d