Explicitly set swift bind_port(s) in .conf files
The swift service is in the process of making it a requirement to
explicitly specify the bind_port for account, container & object
servers in their .conf files. See
https://review.openstack.org/#/c/118200/ for details.

The swift configuration in tripleo relies on default ports being
set by the swift service. Without these defaults swift in tripleo
will not function correctly.

This patch explicitly sets bind_port in the swift .conf files.

The bind_ports specified in the swift .conf files, if required
can be overridden using the tripleo passthrough mechanism.

This change also removes the use of hard coded swift port values
in elements/swift/os-refresh-config/configure.d/73-swift. With this
patch the ports are set in the .conf files and don't need to be
defaulted in 73-swift.

Change-Id: I0a8c3862f23f74e162ee384cf257e6e5930cde01