Work around keepalived split brain
When a host reboots, it's possible for keepalived to come up whilst
the host's network is still being reconfigured (which happens
asynchronously). This can cause the netfilter it puts in place to
fail to see announcements from other members of the cluster,
convincing it to bring up the VIP when it already exists on another
host.

The approach here is crude: every five minutes we cause a reload of
keepalived's configuration, which has the side-effect of making it
reappraise its netfilters and regenerate them.

This is a stopgap measure whilst the underlying causes are bottomed
out.

Change-Id: Ic0397d79465869b78a72faef8653aedee321a3e2
Partial-fix: 1367742