Run django-admin collectstatic for horizon
If other apps are installed alongside Horizon, we need to collect
all of the static artifacts, such as CSS and JS -- which Django has
support for, so make use of it.

As a consequence, set Horizon's ROOT_ROOT to a well-known path, and
create it if needed in install.d.

Change-Id: Icb2f511734ab882dda0b7d00ab73bd6b3cb961f7
Closes-Bug: #1363298