Hash instance-id instead of expecting specific format
In versions of cloud-init prior to 0.7.5,
/var/lib/cloud/data/instance-id was something of the form
i-########.  Our mysql-common element used that ######## bit to set
the mysqld server_id.

In cloud-init 0.7.5, and probably above, the instance-id file is a
uuid instead, which means there is no numeric part for us to use as
the server_id.

To address this, I'm hashing the instance-id and using that
value instead of one pulled from the id itself, which should work
for arbitrary formats of the id.  There is some tiny chance of
hash collisions with this method, but the server_id will allow
2 ^ 32 - 1 different values, so the odds of that happening are
extremely small.

Change-Id: Ia43742bb776ea93b6ed281161c3d362f5b803171