haproxy global maxconn is less than default maxconn
The current global maxconn is hard-coded to 4000 and the default
per-service maxconn is 8000. As a result we have a potential
denial of service attack with a single service using up the
global limit.

This patch makes the global maxconn and the per-service maxconn
configurable. The default global maxconn is left at 4000 and the
default value for the default per-service maxconn is set to a much
lower value. With this change, both per-service default and
individual service maxconn settings are configurable.

Note the default per-service level setting of 150 was chosen on the
basis of haproxy handling 20 services, leading to 200 connections
per service. A value lower than 200 was chosen to allow wriggle room
for specific service maxconn levels to be set higher than others.

Change-Id: I0f1c6bd47f25fb8839c6ebc7fd5f1afb3fabd986