Change initial state and failback policy for VIP
This change is to reduce VIP transition.
The current defaults cause the VIP to automatically
failback to a higher priority controller or one with
a higher IP if priorities are the same ( which they are
by default) This change enables nopreempt by default,
thereby disabling automatic failback when a higher priority
controller becomes available.  It also changes the
default initial state to BACKUP (required for nopreempt).

There are also two new configurables:

   keepalived.state        - BACKUP
   keepalived.preempt      - false

Use keepalived.state=MASTER,keepalived.preempt=true for
original behaviour.

Change-Id: Ie27327286524ed93326ae49c7ca420809fb553a1