Remove redundancies between neutron-router and neutron-network-node
This change removes the redundancies created by the introduction of the
neutron-router as part of enabling Neutron DVR support in tripleo. Now
rather than installing the l3-agent and metadata-agent as part of the
neutron-network-node element, the element instead has a dependency on
the neutron-router element which does the desired agent installations.

Change-Id: Ief61d028fa3b0674b44d8fb94f585209fd7dc635
blueprint: support-neutron-dvr