Parametrize Swift mount_check setting
Currently mount_check is set to false in Swift's account, container
and object -server.conf files.  This is because by default TripleO
Swift does not use mounted devices for data storage.  However
we (HP) and likely others are in the process of deploying TripleO
on servers which use mounted devices for data storage.  To
accommodate this, we propose to set mount_check to the Heat
metadata swift.mount-check if it exists, otherwise mount_check is
set to false.  We will submit corresponding Heat template changes
to set swift.mount-check, with a default value of false.

Change-Id: I6a67c49120b5dfa95d28d71eb7f9049946e9c5aa