Simplify keepalived custom policy
Instead of allowing access to individual types to keepalived, we
now grant keepalived access to read the processes for all types.

This change was suggested in
https://bugzilla.redhat.com/show_bug.cgi?id=1151647

This also makes the custom policy work on both Fedora and RHEL.
The previous custom policy would not install on RHEL 7.0 because
some types were not defined in 7.0.

Change-Id: Ic7adbd14ef27959f0a991127b5213384c9e46be3