Correct the README.md descriptions
Drops the 'from git' from many of the elements. This implies
that the elements always use git... which they don't if
you are using packages.

Also, expands many of the descriptions to say 'and configures' since
most of these elements also contain the OAC/ORC scripts to actually
configure things as well.

Change-Id: I3cbf80c99cfedd73770c9d24e140b7d2443f0dda