Fix horizon package install
We need to copy the local_settings.py file over to the packaged location,
similar to what we do for the source install. Otherwise, config changes to
local_settings.py applied by os-apply-config are not actually in effect. We
must also edit the file to adjust the STATIC_ROOT location.

Add a missing "set -o pipefail" as well.

Change-Id: Ibe0df1596238b5b7f15ac75ce751bc7312b67479