Add pycrypto install to swift-proxy element
We've encountered an issue with memcache encryption of authtoken
entries on Swift proxies where the authtoken middleware errors
out since pycrypto isn't installed.  This has been exposed on
standalone Swift servers.  To fix the issue, we've modified the
swift-proxy element so that pycrypto is installed along with
keystonemiddleware and python-swiftclient.

Change-Id: I727087a85df7ed94fd0ecac86c78826328e644f4
Closes-Bug: #1393790