Set explicit route for VIPs
When VIPs are moved around explicit route for each VIP is
set to make sure that connections opened from local node to VIP
don't use VIP as source address - if VIP is used as source address
and is moved away then connections are stalled.

This explicit routes are removed when VIPs are moved to another node.

Closes-Bug: 1376200
Change-Id: Ie5d55fbe33231519015b313b1f9210ac3614266e