Fix indentation errors in init-neutron-ovs
Prior to this change, init-neutron-ovs would generate an
neutron-openvswitch-agent.conf with various indentation errors. In one
example, something that looked approximately like this was seen:

  - type: ovs-bridge
    members:
      - type: interface
    use_dhcp: true
      - type: vlan
        vlan_id: 1234

which fails on the "type: vlan" line as that's not a legal place to have
a mapping.

I think this change makes this snippet revert to the original intent
here; if I've done it correctly, "- type: vlan" should be at the same
level as "- type: interface ", as part of the "members:" list

Change-Id: I5cea0545928d0bb47c37a6f4fd4c5687e323b99d