Improve rabbitmq_join_cluster on failure.
Add an option for rabbitmq_join_cluster to '--clear-on-error' and
an option for rabbitmq_reset_node to '--force-reset'. This allows
the DB to be cleared down forcefully when a cluster join fails as
the local DB may be in an inconsistent state.

In conjunction with the change in 51-rabbitmq (where the cluster
formation is orchestrated), this allows a node that has failed to
join a cluster to retry more safely.

Finally, after we join cluster, use `update_cluster_nodes` to ensure
the local node is fully synced with the cluster.

Change-Id: Ie53dbab52f0ead23f204dec418056f47480bb20a