Change the kill_metadata executable strings in Neutron
The scripts 80-neutron-dhcp-agent and 80-neutron-router modifies
the neutron rootwrap filter files l3.filters and dhcp.filters.
They replaces the string "/usr/bin/python" with the path to
the python found in the NEUTRON_VENV_DIR/bin directory.

The neutron patch https://review.openstack.org/#/c/118296
replaced "/usr/bin/python" with "python" in the kill_metadata
filters. As a result, the 80-neutron-dhcp-agent and 80-neutron-
router scripts need to be modified to make the same replacement
of "/usr/bin/python" with "python".

If this were not corrected, the neutron L3 agent will get a
rootwrap error (no filters matched) when it tries to kill a
metadata-proxy process and caused the cleanup of a deleted
router to fail.

Closes-Bug: #1398591
Change-Id: Id520ea27f2803447eff654d14ba8cbb388502a52